handled server's response on client's side
